LayoutTests/platform/ios/editing/text-iterator/hidden-textarea-selection-quirk-expected.txt



FAIL Selecting in barely visible textarea assert_equals: expected "test text" but got ""
PASS Selecting in barely visible div 
PASS Selecting in textarea hidden by container 
PASS Selecting in div hidden by container 
FAIL Selecting in absolute positioned textarea hidden by container assert_equals: expected "test text" but got ""
PASS Selecting in absolute positioned div hidden by container 
FAIL Selecting in absolute positioned zero content width textarea hidden by container (quirk behavior) assert_equals: expected "test text" but got ""
PASS Selecting in absolute positioned zero content width div hidden by container 
